Made DAO classes interfaces
